lint code

+++ code/app/src/main/java/com/example/experiment_automata/backend/experiments/MeasurementExperiment.java
@@ -1,12 +1,10 @@
 package com.example.experiment_automata.backend.experiments;
 
 import com.example.experiment_automata.backend.trials.MeasurementTrial;
-import com.example.experiment_automata.backend.trials.Trial;
 import com.github.mikephil.charting.data.BarEntry;
 import com.github.mikephil.charting.data.Entry;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
@@ -132,7 +130,7 @@ public class MeasurementExperiment extends Experiment<MeasurementTrial> {
             }
         }
         // Return sum of all the results divided by the number of trials
-        return (float) sum / getSize();
+        return sum / getSize();
     }
 
     /**
